Extract ModelType isSubtypeOf unit tests
